resources/swe-book/html/tools/requirements.txt
absl-py
beautifulsoup4
